code: bugfix in clean osm data

+++ scripts/clean_osm_data.py
@@ -613,10 +613,10 @@ def fill_circuits(df):
                 for (vc, vf) in zip(row["cables"], row["tag_frequency"])
             ]
         )
-
+    
     df.loc[df_match_by_cables.index, "circuits"] = df_match_by_cables.apply(
         _filter_cables, axis=1
-    )
+    ).astype("float64")
 
     # length of cables elements is larger than frequency; the last cable data are merged to match
     df_merge_by_cables = df[to_fill_merge][["tag_frequency", "cables"]].copy()
